+++ langs/typescript/tests/conformance/form-001-to-010-form-vm.test.ts
@@ -205,6 +205,7 @@ describe("FORM-008", () => {
     });
 
     sut.setModel(makeModel("Bob", 2));
+    messages.length = 0;
     sut.denyCommand.execute();
 
     sub.unsubscribe();